+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Future
 * Sensitivity Calculations
    - Updated mappings
    - Updated display variations
+* Observation page:
+ - Fix effective resolution not updating on Spectral Averaging changes
+ - Fix Spectral Resolution not updating on bandwidth changes for Zoom modes
 * Testing
    - Renamed all component test files to conform to new standards
    - Added standard functions to help simplify testing process